fix: refine route watch restart marker
